added gtk support. Fixes #34650

+++ dev-db/unixODBC/unixODBC-2.2.6.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2003 Gentoo Technologies, Inc.
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/dev-db/unixODBC/unixODBC-2.2.6.ebuild,v 1.1 2003/11/21 19:16:19 rphillips Exp $
+# $Header: /var/cvsroot/gentoo-x86/dev-db/unixODBC/unixODBC-2.2.6.ebuild,v 1.2 2003/12/01 20:25:28 rphillips Exp $
 
 DESCRIPTION="ODBC Interface for Linux"
 HOMEPAGE="http://www.unixodbc.org/"
@@ -9,7 +9,7 @@ SRC_URI="http://www.unixodbc.org/${P}.tar.gz"
 LICENSE="GPL-2"
 SLOT="0"
 KEYWORDS="x86 ~ppc ~hppa ~alpha amd64"
-IUSE="qt"
+IUSE="qt gtk"
 
 DEPEND="virtual/glibc
 	>=sys-libs/readline-4.1
@@ -32,10 +32,32 @@ src_compile() {
 		    ${myconf} || die
 
 	make || die
+
+	if [ "`use gtk`" ]
+	then
+		cd gODBCConfig
+		./configure --host=${CHOST} \
+				--prefix=/usr \
+				--sysconfdir=/etc/unixODBC \
+				${myconf} || die
+
+		# not sure why these symlinks are needed. busted configure, i guess...
+		ln -s ../depcomp .
+		ln -s ../libtool .
+		make || die
+		cd ..
+	fi
 }
 
 src_install() {
 	make DESTDIR=${D} install || die
+
+	if [ "`use gtk`" ]
+	then
+		cd gODBCConfig
+		make DESTDIR=${D} install || die
+		cd ..
+	fi
 
 	dodoc AUTHORS COPYING ChangeLog NEWS README*
 	find doc/ -name "Makefile*" -exec rm '{}' \;
